extract syncIssues action
